+++ TravelDiscovery/Models/Place.swift
@@ -1,0 +1,12 @@
+//
+//  Place.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 06/06/23.
+//
+
+import Foundation
+
+struct Place: Decodable, Hashable {
+    let name, thumbnail: String
+}
